+++ trakker/src/main/java/com/example/trakker/model/trip/dao/TripDAO.java
@@ -1,13 +1,10 @@
 package com.example.trakker.model.trip.dao;
 
 
-import com.example.trakker.model.faq.dto.FaqDTO;
 import com.example.trakker.model.trip.dto.TripDTO;
-import com.example.trakker.utils.ItemSearchVO;
 import com.example.trakker.utils.PagingInfoVO;
 import com.example.trakker.utils.ResponseResultList;
 
-import javax.servlet.http.HttpSession;
 import java.util.List;
 import java.util.Map;
 
@@ -15,8 +12,8 @@ public interface TripDAO {
     List<TripDTO> list() throws Exception;
     void insert(TripDTO dto) throws Exception;
     void update(TripDTO dto) throws Exception;
-    void delete(int t_num) throws Exception;
-    TripDTO view(int t_num) throws Exception;
+    void delete(long t_num) throws Exception;
+    TripDTO view(long t_num) throws Exception;
     Integer count(Map data);
     List<TripDTO> listPage(Map data);
 }
