+++ trakker/src/main/java/com/example/trakker/model/trip/dao/TripDAOImpl.java
@@ -1,16 +1,10 @@
 package com.example.trakker.model.trip.dao;
 
-import com.example.trakker.model.faq.dto.FaqDTO;
+
 import com.example.trakker.model.trip.dto.TripDTO;
-import com.example.trakker.utils.ItemSearchVO;
-import com.example.trakker.utils.PagingInfoVO;
-import com.example.trakker.utils.ResponseResultList;
 import org.apache.ibatis.session.SqlSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
-
-import javax.servlet.http.HttpSession;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
@@ -36,19 +30,19 @@ public class TripDAOImpl implements TripDAO {
     }
 
     @Override
-    public void delete(int t_num) throws Exception {
+    public void delete(long t_num) throws Exception {
         sqlSession.delete("trip.delete", t_num);
     }
 
 
     @Override
-    public TripDTO view(int t_num) throws Exception {
+    public TripDTO view(long t_num) throws Exception {
         return sqlSession.selectOne("trip.view", t_num);
     }
 
     @Override
     public Integer count(Map data) {
-      return sqlSession.selectOne("trip.listPageCount", data);
+        return sqlSession.selectOne("trip.listPageCount", data);
     }
 
     @Override
@@ -56,3 +50,4 @@ public class TripDAOImpl implements TripDAO {
         return sqlSession.selectList("trip.listPage", data);
     }
 }
+
